+++ client/.prettierignore
@@ -1,1 +1,1 @@
-SwaggerApi.ts
+ImisSwaggerApi.ts
