+++ client/README.md
@@ -44,7 +44,7 @@ You can also use `yarn lint --fix` to resolve most lint errors
 ### Server communication
 We communicate with the server using REST. The documentation is generated by swagger.
 We are using the `swagger-typescript-api` package to automatically generate an API client as part
- of the build process (`@/api/SwaggerApi.ts`). You can re-generate this file by issuing
+ of the build process (`@/api/ImisSwaggerApi.ts`). You can re-generate this file by issuing
   `./gradlew generateClient`. Do not edit this file manually, as changes will be overwritten during
    regeneration!
