+++ client/src/util/helper-functions.ts
@@ -1,5 +1,5 @@
 import moment from 'moment'
-import { Patient, InstitutionImpl } from '../api/SwaggerApi'
+import { Patient, InstitutionImpl } from '../api/ImisSwaggerApi'
 
 export function getDate(date: string) {
   if (date !== undefined && date !== null && date !== 'null')
